create cmdutils.go
